Split checkBucket into utils file
