+++ s3utils.py
@@ -1,4 +1,5 @@
 import sh
+import requests
 
 
 def getBucketSize(bucketName):
@@ -8,3 +9,39 @@ def getBucketSize(bucketName):
 
     # Get the last line of the output, get everything to the right of the colon, and strip whitespace
     return a.splitlines()[len(a.splitlines())-1].split(":")[1].strip()
+
+
+def checkBucket(bucketName, region):
+    """ Does a simple GET request with the Requests library and interprets the results.
+
+    site - A domain name without protocol (http[s])
+    region - An s3 region. See: https://docs.aws.amazon.com/general/latest/gr/rande.html#s3_region
+    """
+
+    # Concat domain name with the default region
+    bucketDomain = 'http://' + bucketName + '.s3-' + region + '.amazonaws.com'
+    try:
+        r = requests.get(bucketDomain)
+    except requests.exceptions.ConnectionError:
+        # Couldn't resolve the hostname. Definitely not a bucket.
+        message = "{0:>16} : {1}".format("[not found]", bucketName)
+        return 900, message
+    if r.status_code == 200:
+        # Successfully found a bucket!
+        message = "{0:<7}{1:>9} : {2}".format("[found]", "[open]", bucketName + ":" + region + " - " +
+                                              getBucketSize(bucketName))
+        return 200, message
+    elif r.status_code == 301:
+        # We got the region wrong. The 'x-amz-bucket-region' header will give us the correct one.
+        return 301, r.headers['x-amz-bucket-region']
+    elif r.status_code == 403:
+        # We probably need to have an AWS account defined.
+        message = "{0:>15} : {1}".format("[found] [closed]", bucketName + ":" + region)
+        return 403, message
+    elif r.status_code == 404:
+        # This is definitely not a valid bucket name.
+        message = "{0:>15} : {1}".format("[not found]", bucketName)
+        return 404, message
+    else:
+        raise ValueError("Got an unhandled status code back: " + str(r.status_code) + " for site: " + bucketName + ":" + region)
+
